# backend-laravel-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Receptarium  API",
        "contact": {
            "name": "David Pérez Romero",
            "email": "contacto@david-pr.com"
        },
        "license": {
            "name": "Licencia",
            "url": "http://www.example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/recetas-categoria/{id}": {
            "get": {
                "tags": [
                    "Receta Controller"
                ],
                "summary": "Show a recipe in categori",
                "description": "Show a recipes in categories",
                "operationId": "c9994c9f5f1f8df215509a08b2e1eeaa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": " recipes in categories"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/receta": {
            "post": {
                "tags": [
                    "Receta Controller"
                ],
                "summary": "Store a recipe",
                "description": "Store a recipe",
                "operationId": "4b2d71645248f7dff1802c4c9aed1538",
                "requestBody": {
                    "description": "Recipe data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecetaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recipe stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/receta/{id}": {
            "get": {
                "tags": [
                    "Receta Controller"
                ],
                "summary": "Show a recipe by ID",
                "description": "Show a recipe",
                "operationId": "b2a7891649b5dc38718395c9012d0c0d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recipe found"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Receta Controller"
                ],
                "summary": "Update recipe",
                "description": "Update  recipe",
                "operationId": "78be13e5a44ee0948463d97ae49cf701",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Recipe data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RecetaRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recipe stored",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "ok"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Receta Controller"
                ],
                "summary": "Delete a recipe by ID",
                "description": "Delete a recipe",
                "operationId": "ce10243da9ea4f715e3800e880f2060f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the recipe",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "203": {
                        "description": "Deleted recipe"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "RecetaRequest": {
                "properties": {
                    "titulo": {
                        "type": "string",
                        "example": "Test receta"
                    },
                    "ingredientes": {
                        "type": "string",
                        "example": "ingrediente 1, ingrediente 2"
                    },
                    "tiempo_preparacion": {
                        "type": "string",
                        "example": "20 m"
                    },
                    "num_personas": {
                        "type": "integer",
                        "example": 1
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Descripcion de la receta"
                    },
                    "categorias_recetas_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "user_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}